make <subscribe> a component
